requirements.txt
kokoro>=0.9.0
numpy
scipy
soundfile
pyyaml
tqdm
jupyter
ipykernel
